refactor for add food handler and fixed buggy daystreak/goldencarrot update
