+++ components/SearchFoodsForm/styles.module.css
@@ -39,22 +39,6 @@
   height: 185px;
 }
 
-.buttons {
-  display: flex;
-  gap: 11px;
-}
-
-.addButton {
-  background-color: var(--purple-light);
-  height: 50px;
-  width: 148px;
-  border-radius: 4px;
-  border: none;
-  margin: 0;
-  color: white;
-  font-size: medium;
-}
-
 @media (min-width: 768px) {
   .container > input {
     height: 92px;
@@ -65,14 +49,5 @@
   }
   .backgroundImage > img {
     height: 280px;
-  }
-  .buttons {
-    gap: 40px;
-  }
-  .addButton {
-    line-height: 92px;
-    font-size: 24px;
-    height: 92px;
-    width: 278px;
   }
 }
